# docs/static/api/openapi.yaml
openapi: 3.0.3
info:
  title: ChatAPI
  version: 1.0.0
  description: |
    Self-hosted chat infrastructure for AI-powered apps. Single-tenant per deployment.

    ## Authentication

    All endpoints except `/health` and `/metrics` require a JWT Bearer token:

    ```
    Authorization: Bearer <jwt>
    ```

    Your backend signs JWTs with `JWT_SECRET`. The `sub` claim is the user ID.

    WebSocket connections accept the token as a query parameter:
    ```
    ws://your-host/ws?token=<jwt>
    ```

  license:
    name: MIT
    url: https://github.com/hastenr/chatapi/blob/main/LICENSE
  contact:
    url: https://github.com/hastenr/chatapi

servers:
  - url: http://localhost:8080
    description: Local development

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    RoomID:
      name: room_id
      in: path
      required: true
      schema:
        type: string
    MessageID:
      name: message_id
      in: path
      required: true
      schema:
        type: string
    BotID:
      name: bot_id
      in: path
      required: true
      schema:
        type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Not the resource owner
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]

    Room:
      type: object
      properties:
        room_id:
          type: string
        type:
          type: string
          enum: [dm, group]
        name:
          type: string
        last_seq:
          type: integer
        metadata:
          type: string
          description: Arbitrary JSON for app-level context
        created_at:
          type: string
          format: date-time
      required: [room_id, type, last_seq, created_at]

    RoomMember:
      type: object
      properties:
        chatroom_id:
          type: string
        user_id:
          type: string
        role:
          type: string
        joined_at:
          type: string
          format: date-time
      required: [chatroom_id, user_id, role, joined_at]

    Message:
      type: object
      properties:
        message_id:
          type: string
        chatroom_id:
          type: string
        sender_id:
          type: string
        seq:
          type: integer
        content:
          type: string
        meta:
          type: string
          description: Arbitrary JSON metadata
        created_at:
          type: string
          format: date-time
      required: [message_id, chatroom_id, sender_id, seq, content, created_at]

    Bot:
      type: object
      properties:
        bot_id:
          type: string
        name:
          type: string
        mode:
          type: string
          enum: [llm, external]
        provider:
          type: string
          enum: [openai, anthropic]
        base_url:
          type: string
        model:
          type: string
        system_prompt:
          type: string
        max_context:
          type: integer
        created_at:
          type: string
          format: date-time
      required: [bot_id, name, mode, max_context, created_at]

    CreateBotRequest:
      type: object
      properties:
        name:
          type: string
        mode:
          type: string
          enum: [llm, external]
        provider:
          type: string
          enum: [openai, anthropic]
        base_url:
          type: string
          description: Optional override for OpenAI-compatible endpoints
        model:
          type: string
        api_key:
          type: string
        system_prompt:
          type: string
        max_context:
          type: integer
          description: Max messages to include in context window (default 20)
      required: [name, mode]

    UndeliveredMessage:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
        chatroom_id:
          type: string
        message_id:
          type: string
        seq:
          type: integer
        attempts:
          type: integer
        created_at:
          type: string
          format: date-time
        last_attempt_at:
          type: string
          format: date-time
      required: [id, user_id, chatroom_id, message_id, seq, attempts, created_at]

paths:
  /health:
    get:
      summary: Health check
      security: []
      tags: [System]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /metrics:
    get:
      summary: Live server counters
      security: []
      tags: [System]
      responses:
        "200":
          description: Server metrics
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true

  /rooms:
    get:
      summary: List rooms the authenticated user belongs to
      tags: [Rooms]
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Create a room
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [dm, group]
                members:
                  type: array
                  items:
                    type: string
                  description: User IDs to add (including the caller)
                name:
                  type: string
                metadata:
                  type: string
                  description: Arbitrary JSON
              required: [type, members]
      responses:
        "200":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /rooms/{room_id}:
    get:
      summary: Get room details
      tags: [Rooms]
      parameters:
        - $ref: "#/components/parameters/RoomID"
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update room name or metadata
      tags: [Rooms]
      parameters:
        - $ref: "#/components/parameters/RoomID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                metadata:
                  type: string
                  description: Arbitrary JSON
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /rooms/{room_id}/members:
    get:
      summary: List room members
      tags: [Rooms]
      parameters:
        - $ref: "#/components/parameters/RoomID"
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      summary: Add a member (user or bot) to a room
      tags: [Rooms]
      parameters:
        - $ref: "#/components/parameters/RoomID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: User ID or bot ID to add
              required: [user_id]
      responses:
        "200":
          description: Member added
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /rooms/{room_id}/messages:
    get:
      summary: List messages in a room (paginated)
      tags: [Messages]
      parameters:
        - $ref: "#/components/parameters/RoomID"
        - name: after_seq
          in: query
          schema:
            type: integer
          description: Return messages with seq > after_seq
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      summary: Send a message to a room
      tags: [Messages]
      parameters:
        - $ref: "#/components/parameters/RoomID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                meta:
                  type: string
                  description: Arbitrary JSON metadata
              required: [content]
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /rooms/{room_id}/messages/{message_id}:
    put:
      summary: Edit a message (sender only)
      tags: [Messages]
      parameters:
        - $ref: "#/components/parameters/RoomID"
        - $ref: "#/components/parameters/MessageID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required: [content]
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete a message (sender only)
      tags: [Messages]
      parameters:
        - $ref: "#/components/parameters/RoomID"
        - $ref: "#/components/parameters/MessageID"
      responses:
        "204":
          description: Message deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /acks:
    post:
      summary: Acknowledge message delivery up to a sequence number
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                room_id:
                  type: string
                seq:
                  type: integer
              required: [room_id, seq]
      responses:
        "200":
          description: Acknowledged
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /bots:
    get:
      summary: List registered bots
      tags: [Bots]
      responses:
        "200":
          description: List of bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Register a new bot
      tags: [Bots]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBotRequest"
      responses:
        "200":
          description: Bot registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /bots/{bot_id}:
    get:
      summary: Get a bot by ID
      tags: [Bots]
      parameters:
        - $ref: "#/components/parameters/BotID"
      responses:
        "200":
          description: Bot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete a bot
      tags: [Bots]
      parameters:
        - $ref: "#/components/parameters/BotID"
      responses:
        "204":
          description: Bot deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /admin/dead-letters:
    get:
      summary: List failed (dead-letter) messages and notifications
      tags: [Admin]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        "200":
          description: Dead-letter items
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/UndeliveredMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"